third_party/blink/web_tests/external/wpt/sanitizer-api/element-set-sanitized-html.https-expected.txt

This is a testharness.js-based test.
Found 85 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Sanitizer: <template>.setHTML("<em>Hello</em>", ...) obeys parse context.
  assert_equals: expected "<em>Hello</em>" but got ""
[FAIL] Sanitizer: <template>.setHTML("<td>data</td>", ...) obeys parse context.
  assert_equals: expected "<td>data</td>" but got ""
[FAIL] Sanitizer: Element.setHTML with config: string
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: html fragment
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: broken html
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: empty object
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: number
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: zeros
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: arithmetic
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: empty string
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: undefined
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: document
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: html without close tag
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: scripts for default configs
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: script not as root
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: script deeper in the tree
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: onclick scripts
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: plaintext
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: xmp
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: invalid config_input
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: empty removeElements list
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: test html without close tag with removeElements list ['div']
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: default behavior for custom elements
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: allow custom elements
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: allow custom elements with allow elements
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: disallow custom elements
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: allow custom elements with drop list contains ["custom-element"]
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: test script with ["script"] as removeElements list
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: removeElements list ["test-element", "i"]}
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: removeElements list ["dl", "p"]}
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: elements list ["p"]
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: elements list has no influence to removeElements
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: empty removeAttributes list with id attribute
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: removeAttributes list ["id"] with id attribute
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: removeAttributes list ["data-attribute-with-dashes"] with dom dataset js access
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: elements list with <p> attributes: ["title"] and div attributes: ["id"] lists
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: elements list with <p> removeAttributes: ["title"]  and div removeAttributes: ["id"] lists
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: elements list with <div> attributes: ["id"] and removeAttributes: ["id"] lists
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: elements list with <div> attributes: ["id", "title"] does not override empty attributes: [] list
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: elements list with <div> attributes: ["id", "title"] does not override removeAttributes: ["id", "title"] list
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: elements list with <div> removeAttributes: ["id", "title"] is effective even with attributes: ["id", "title"] list
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: attributes list ["id"] with id attribute and onclick scripts
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: attributes list has no influence to removeAttributes list
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: Template element
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: HTMLAnchorElement with javascript protocal
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: HTMLAnchorElement with javascript protocal start with space
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: HTMLAnchorElement
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: HTMLAreaElement with javascript protocal
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: HTMLAreaElement with javascript protocal start with space
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: HTMLAreaElement
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: HTMLFormElement with javascript action
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: HTMLFormElement with javascript action start with space
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: HTMLFormElement
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: HTMLInputElement with javascript formaction
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: HTMLInputElement with javascript formaction start with space
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: HTMLInputElement
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: HTMLButtonElement with javascript formaction
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: HTMLButtonElement with javascript formaction start with space
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: HTMLButtonElement
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: malformed HTML
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: HTML with comments; comments not allowed
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: HTML with comments; comments
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: HTML with comments; !comments
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: HTML with comments deeper in the tree
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: HTML with comments deeper in the tree, comments
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: HTML with comments deeper in the tree, !comments
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: Unknown HTML names (HTMLUnknownElement instances) should not match elements parsed as non-HTML namespaces.
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: Unknown HTML names (HTMLUnknownElement instances) should not match elements parsed as non-HTML namespaces when nested.
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: removeElements list ["I", "DL"]}
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: removeElements list ["i", "dl"]}
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: removeElements list ["i", "dl"]} with uppercase HTML
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: removeAttributes list ["ID"] with id attribute
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: removeAttributes list ["ID"] with ID attribute
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: removeAttributes list ["id"] with ID attribute
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: removeElements with unknown elements and without unknownMarkup
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: replaceWithChildrenElements with unknown elements and without unknownMarkup
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: elements with unknown elements and without unknownMarkup
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: removeElements with unknown elements and with unknownMarkup
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: replaceWithChildrenElements with unknown elements and with unknownMarkup
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: elements with unknown elements and with unknownMarkup
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: attributes: unknown attributes and without unknownMarkup
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: attributes: unknown attributes and with unknownMarkup
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: removeAttributes: unknown attributes and without unknownMarkup
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML with config: removeAttributes unknown attributes and with allowUnknownMarkup
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
[FAIL] Sanitizer: Element.setHTML options dictionary #2 uses default.
  Failed to execute 'setHTML' on 'Element': Failed to read the 'sanitizer' property from 'ElementSetHTMLOptions': Failed to convert value to 'Sanitizer'.
Harness: the test ran to completion.
